Producer/requirements.txt
pymongo
schedule
kafka-python
